+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to BPM-Caddy will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.216.0] - 2026-09-17
 
 ### Added
 - **Et la décroissance plasmatique pour tout ce qui n'est pas une